mod: PR Review 반영 - Bottom Sheet State는Sheet 컴포넌트 내에서 관리
